requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
asgiref==3.7.2
    # via
    #   -r requirements/base.txt
    #   django
click==8.1.7
    # via code-annotations
code-annotations==1.5.0
    # via -r requirements/test.in
coverage[toml]==7.3.2
    # via pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   openedx-filters
exceptiongroup==1.1.3
    # via pytest
iniconfig==2.0.0
    # via pytest
jinja2==3.1.2
    # via code-annotations
markupsafe==2.1.3
    # via jinja2
openedx-filters==1.6.0
    # via -r requirements/base.txt
packaging==23.2
    # via pytest
pbr==5.11.1
    # via stevedore
pluggy==1.3.0
    # via pytest
pytest==7.4.2
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==4.1.0
    # via -r requirements/test.in
pytest-django==4.5.2
    # via -r requirements/test.in
python-slugify==8.0.1
    # via code-annotations
pytz==2023.3.post1
    # via
    #   -r requirements/base.txt
    #   django
pyyaml==6.0.1
    # via code-annotations
sqlparse==0.4.4
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.1.0
    # via code-annotations
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   pytest
typing-extensions==4.8.0
    # via
    #   -r requirements/base.txt
    #   asgiref
web-fragments==2.1.0
    # via -r requirements/base.txt
